refactor(viewer): extract nodal and element field painting from the FEA loader
paintNodeField and paintElemField (fea/streaming) each take the session
handle, the field + step, the user's view settings, the resolved warp
source and the blob fetchers, and paint every surface that carries the
field with the same reduction, colormap, contour and morph influence.
The loader keeps the shared prelude -- reduction, warp resolution --
and routes on per_type.

warpInfluence.test reads the loader's source text to pin that every
paint passes the influence it was asked for; it now reads the loader
together with its fea/streaming package, where the paints live.
